Fixes to stop PHP process from hanging
Under certain circumstances, mod_suphp spawned PHP processes would hang. I suspect it has to do with the connection being interrupted. Because the process would hang, and apparently because this module was blocking with a wait() call, the two would sit deadlocked. This leads to several "running" processes that would never complete, which could eventually eat RAM up.

log_cpu_elapsed() now performs a non-blocking wait and checks to make sure the corresponding conn_rec's aborted field is zero. It will continue if the child process it is waiting on finishes, or the connection is aborted.